move ui-utils to new utils dir
